moved vello to separate example
