--- crates/gosub_useragent/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "gosub_useragent"
-version = "0.1.0"
-edition = "2021"
-authors = ["Gosub Community <info@gosub.io>"]
-license = "MIT"
-
-[dependencies]
-gosub_shared = { path = "../gosub_shared" }
-gosub_renderer = { path = "../gosub_renderer" }
-winit = "0.30.5"
-slotmap = "1.0.7"
-log = "0.4.22"
-anyhow = "1.0.94"
-url = "2.5.4"
-image = "0.25.5"
-
-
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-web-sys = "0.3.72"
-wasm-bindgen-futures = "0.4.47"
