--- crates/gosub_useragent/src/lib.rs
@@ -1,6 +1,0 @@
-pub mod application;
-pub mod event_loop;
-pub mod tabs;
-pub mod window;
-
-pub use winit;
