--- examples/gtk-renderer/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "gtk-renderer"
-version = "0.1.0"
-edition = "2021"
-authors = ["Gosub Community <info@gosub.io>"]
-description = "Simple renderer using GTK4 and the cairo backend rendering system"
-license = "MIT"
-
-[dependencies]
-gosub_shared = { path = "../../crates/gosub_shared", features = [] }
-gosub_html5 = { path = "../../crates/gosub_html5", features = [] }
-gosub_css3 = { path = "../../crates/gosub_css3", features = [] }
-gosub_taffy = { path = "../../crates/gosub_taffy", features = [] }
-gosub_renderer = { path = "../../crates/gosub_renderer", features = [] }
-gosub_rendering = { path = "../../crates/gosub_render_utils", features = [] }
-gosub_cairo = { path = "../../crates/gosub_cairo", features = [] }
-gtk4 = "0.9.4"
-url = "2.5.4"
-log = "0.4.22"
-simple_logger = "5.0.0"
-clap = "4.5.23"
