+++ examples/vello-renderer/README.md
@@ -1,0 +1,9 @@
+# Vello Renderer Example
+
+This example demonstrates how to use a winit window to render a website.
+
+## Usage
+
+```bash
+cargo run --release --package vello-renderer -- <URL>
+```
